Adding _count handler. Adding testutils to mock elasticsearch responses.

+++ rubber/resource.py
@@ -1,5 +1,6 @@
 import requests
 from rubber import settings
+from rubber.testutils import ResponseMock
 from instanceutils import data_to_json
 
 class Resource(object):
@@ -22,15 +23,23 @@ class Resource(object):
         return response
 
     def get(self, data=None, **kwargs):
+        if settings.RUBBER_MOCK_HTTP_RESPONSE:
+            return self.wrapper(ResponseMock(settings.RUBBER_MOCK_HTTP_RESPONSE))
         return self.wrapper(requests.get(self.base_url + self.path, data=data_to_json(data), **kwargs))
 
     def put(self, data=None, **kwargs):
+        if settings.RUBBER_MOCK_HTTP_RESPONSE:
+            return self.wrapper(ResponseMock(settings.RUBBER_MOCK_HTTP_RESPONSE))
         return self.wrapper(requests.put(self.base_url + self.path, data=data_to_json(data), **kwargs))
 
     def post(self, data=None, **kwargs):
+        if settings.RUBBER_MOCK_HTTP_RESPONSE:
+            return self.wrapper(ResponseMock(settings.RUBBER_MOCK_HTTP_RESPONSE))
         return self.wrapper(requests.post(self.base_url + self.path, data=data_to_json(data), **kwargs))
 
     def delete(self, data=None, **kwargs):
+        if settings.RUBBER_MOCK_HTTP_RESPONSE:
+            return self.wrapper(ResponseMock(settings.RUBBER_MOCK_HTTP_RESPONSE))
         return self.wrapper(requests.delete(self.base_url + self.path, data=data_to_json(data), **kwargs))
 
     def __call__(self, *args, **kwargs):
